requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Caching==2.1.0
Flask-Assets==2.1.0
email-validator==2.1.0.post1
python-dotenv==1.0.0
openai==1.12.0
httpx>=0.25.0
requests==2.31.0
Faker==22.5.1
colorlog==6.8.0
gunicorn==21.2.0
cssmin==0.2.0
jsmin==3.0.1
SQLAlchemy>=2.0.36
WTForms==3.1.1
openpyxl>=3.0.0
Pillow>=10.0.0

# 生产环境数据库驱动（可选）
psycopg2-binary>=2.9.9  # PostgreSQL
# mysqlclient>=2.2.0    # MySQL（需要时取消注释）